+++ app/assets/javascripts/mobile/tag_following.js
@@ -1,7 +1,8 @@
 $(document).ready(function(){
   $(".tag_following_action").bind("tap click", function(evt){
     evt.preventDefault();
-    var button = $(this),
+    var tagFollowing,
+        button = $(this),
         tagName = button.data("name");
 
     if(button.hasClass("btn-success")){
@@ -19,11 +20,11 @@ $(document).ready(function(){
         button.removeClass("btn-success").addClass("btn-danger");
         button.text(Diaspora.I18n.t("stream.tags.stop_following", {tag: tagName}));
       }).fail(function() {
-        alert(Diaspora.I18n.t("stream.tags.follow_error", {name: "#" + tagName}));
+        alert(Diaspora.I18n.t("stream.tags.follow_error", {tag: tagName}));
       });
     }
     else if(button.hasClass("btn-danger")){
-      var tagFollowing = _.findWhere(gon.preloads.tagFollowings,{name: tagName});
+      tagFollowing = _.findWhere(gon.preloads.tagFollowings, {name: tagName});
       if(!tagFollowing) { return; }
       $.ajax({
         url: Routes.tagFollowing(tagFollowing.id),
@@ -36,7 +37,28 @@ $(document).ready(function(){
         button.removeClass("btn-danger").addClass("btn-success");
         button.text(Diaspora.I18n.t("stream.tags.follow", {tag: tagName}));
       }).fail(function() {
-        alert(Diaspora.I18n.t("stream.tags.stop_following_error", {name: "#" + tagName}));
+        alert(Diaspora.I18n.t("stream.tags.stop_following_error", {tag: tagName}));
+      });
+    }
+    else if(button.hasClass("only-delete")){
+      tagFollowing = _.findWhere(gon.preloads.tagFollowings, {name: tagName});
+      if(!tagFollowing ||
+        !confirm(Diaspora.I18n.t("stream.tags.stop_following_confirm", {tag: tagName}))) { return; }
+
+      $.ajax({
+        url: Routes.tagFollowing(tagFollowing.id),
+        dataType: "json",
+        type: "DELETE",
+        headers: {
+          "Accept": "application/json, text/javascript, */*; q=0.01"
+        }
+      }).done(function() {
+        button.closest("li").remove();
+        if($("ul.followed_tags li").length === 0){
+          $(".well").removeClass("hidden");
+        }
+      }).fail(function() {
+        alert(Diaspora.I18n.t("stream.tags.stop_following_error", {tag: tagName}));
       });
     }
   });