+++ Changelog.md
@@ -35,6 +35,7 @@
 * Gracefully handle missing `og:url`s [#5926](https://github.com/diaspora/diaspora/pull/5926)
 * Remove private post content from "also commented" mails [#5931](https://github.com/diaspora/diaspora/pull/5931)
 * Add a button to follow/unfollow tags to the mobile interface [#5941](https://github.com/diaspora/diaspora/pull/5941)
+* Add a "Manage followed tags" page to mass unfollow tags in the mobile interface [#5945](https://github.com/diaspora/diaspora/pull/5945)
 
 # 0.5.0.1